[Refactor] move ActionBuilder in core instead of utils
