+++ src/core/ActionRecognition.cpp
@@ -1,6 +1,6 @@
 #include <set>
 #include "procedural/core/Types/ActionType.h"
-#include "procedural/utils/ActionRecognition.h"
+#include "procedural/core/ActionRecognition.h"
 #include <iostream>
 
 namespace procedural {
